// src/@core/enums/message-type.enum.ts
export enum MESSAGES_TYPES {
  SELECT_ELEMENTS = 'SELECT_ELEMENTS',
  RESET_SETTING = 'RESET_SETTING',
  SELECT_PAGINATION = 'SELECT_PAGINATION',
  SHOW_ALERT = 'SHOW_ALERT',
  SEARCH_FOR_TABLE = 'SEARCH_FOR_TABLE',
  TABLE_EXIST = 'TABLE_EXIST',
  SELECT_PAGINATION_STOP = 'SELECT_PAGINATION_STOP',
  DOWNLOAD_TABLE_DATA = 'DOWNLOAD_TABLE_DATA',
  DOWNLOAD_DATA = 'DOWNLOAD_DATA',
  START_SCRAPING = 'START_SCRAPING',
  STOP_SCRAPING = 'STOP_SCRAPING',
  IS_SCRAPING = 'IS_SCRAPING',
  HAS_SAVED_DATA = 'HAS_SAVED_DATA',
  IS_CONTENT_SCRIPT_LOADED = 'IS_CONTENT_SCRIPT_LOADED'
}